+++ package.json
@@ -5,13 +5,20 @@
     "main": "bot.js",
     "author": "Krishna Chaitanya Addanki",
     "dependencies": {
+        "axios": "^0.26.0",
         "cron": "^1.8.2",
         "discord.js": "^13.6.0",
+        "fs-extra": "^10.0.1",
         "rss-parser": "^3.12.0"
     },
     "devDependencies": {
         "eslint": "^8.9.0",
         "eslint-config-airbnb-base": "^15.0.0",
         "eslint-plugin-import": "^2.25.4"
+    },
+    "type": "module",
+    "scripts": {
+        "robot": "node src/robot.js",
+        "bot": "node src/bot.js"
     }
 }
